pyproject.toml
[build-system]
requires = ["maturin>=1.8,<2.0"]
build-backend = "maturin"

[project]
name = "eqloader"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dynamic = ["version"]
[project.optional-dependencies]
dev = [
    "pytest",
    "fake-bpy-module",
    "maturin"
]
blender-dev = ['fake-bpy-module']
[tool.maturin]
python-source = "python"
features = ["pyo3/extension-module"]
